EASY-2502: restoring/refactoring of tests - phase 1 SubmitterSpec (#203)
* new submitter tests, moved ignored ones to SubmitJobSpec
* file-limit argument from Submitter to Mailer
