+++ src/main/scala/nl.knaw.dans.easy.deposit/Submitter.scala
@@ -37,7 +37,6 @@ import scala.util.{ Failure, Success, Try }
 class Submitter(submitToBaseDir: File,
                 groupPrincipal: GroupPrincipal,
                 depositUiURL: String,
-                fileLimit: Int = 200,
                 jobQueue: JobQueueManager,
                 mailer: Mailer,
                 agreementGenerator: AgreementGenerator,
@@ -81,7 +80,6 @@ class Submitter(submitToBaseDir: File,
           draftDepositId = draftDeposit.id,
           depositUiURL = depositUiURL,
           groupPrincipal = groupPrincipal,
-          fileLimit = fileLimit,
           draftBag = draftBag,
           stagedDepositDir = stagedDir,
           submitDir = submitDir,
